requirements.txt
aniso8601==9.0.1
anyio==3.6.2
blinker==1.6.2
cffi==1.15.1
click==8.1.3
cryptography==40.0.2
greenlet==2.0.2
h11==0.14.0
httptools==0.5.0
idna==3.4
importlib-metadata==6.6.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
marshmallow==3.19.0
packaging==23.1
pycparser==2.21
pydantic==1.10.7
PyJWT==2.7.0
python-dotenv==1.0.0
pytz==2023.3
PyYAML==6.0
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.13
starlette==0.27.0
typing-extensions==4.5.0
uvicorn==0.22.0
uvloop==0.17.0
watchfiles==0.19.0
websockets==11.0.3
Werkzeug==2.3.4
zipp==3.15.0
